Add door and hatch actors
Doors and hatches can be opened and closed. Hatches have additional
functionality, such as a way to spawn an actor, which will cause the
hatch to automatically open and bring the actor to the surface.

Add a playground map with external actors enabled to serve as a, well,
playground for experimentation. An example of the door and hatch are
included for demonstration.

The hatches can be extended with a parameter for how long they stay up
or if they should stay up forever. This can be useful to spawn things
like random equipment or power-ups at random spots in the arena.

Note that currently this doesn't work with the arena because we can't
punch holes through the ground. This could potentially be solved by
creating the arena floor from modular tiles, where the hatch can be one
tile. PCG should be able to achieve this.

These are very rough models for now and should be refined.