Add PCG arena
The BP_Arena actor can be placed in a level and configured to generate
an arena using the specified radius, number of spline points, etc. The
generated arena contains a floor with occasional hatches spread
throughout. A high wall surrounds the entire arena so players cannot
escape.

Add a few more FBX files for models. Change the size of the hatch to be
400x400 to match the floor mesh. Use rock models to add some variety to
the arena.

Enable the PCG plugin for use in the BP_Arena actor.